// src/types/api.types.ts
export interface ApiError {
  message: string
  statusCode: number
  error?: string
  details?: unknown
}